Fix PDF/A support by always appending a newline to a stream
The standard PDF spec allows multiple representations for streams that end in a newline. Either ending in a single newline (which serves double purpose as a byte in the stream and a separator for the endstream token), or two newlines. This can be parsed unambiguously based on the provided length.

However, PDF/A requires two newlines if the stream ends in a newline. This change implements this.
PDF/A-1 reference: 6.1.7, pp. 6
PDF reference: Table 3.4, Entries common to all stream dictionaries, pp. 38
Issue thread: https://sourceforge.net/p/naps2/discussion/general/thread/ca12bbfc/